move log, save params
